修改了urlListBean size 变量数据类型应该是Double

+++ app/src/main/java/com/jennifer/andy/simpleeyes/net/entity/AndyInfo.kt
@@ -168,7 +168,7 @@ data class PlayInfoBean(var height: Int,
 
 data class UrlListBean(var name: String,
                        var url: String,
-                       var size: Int) : Serializable
+                       var size: Double) : Serializable
 
 
 data class Category(var categoryInfo: CategoryInfo,
